Changed the layout of the page and updated the CSS files. Also edited the structure of the components files.
