+++ src/App.js
@@ -1,9 +1,10 @@
 import React from 'react';
-import Header from './components/Header';
-import Footer from './components/Footer';
+import Header from './components/Header/Header';
+import Footer from './components/Footer/Footer';
 import Services from './components/Services';
 import Projects from './components/Projects';
 import Contact from './components/Contact';
+import Process from './components/Process';
 import './App.css';
 
 function App() {
@@ -13,6 +14,7 @@ function App() {
             <main>
                 <Services />
                 <Projects />
+                <Process /> {/* Now included in the main section */}
                 <Contact />
             </main>
             <Footer />
