--- src/components/Footer.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Footer = () => {
-    return (
-        <footer>
-            <p>&copy; 2024 My Construction Company. All rights reserved.</p>
-        </footer>
-    );
-}
-
-export default Footer;
