--- src/components/Header.js
@@ -1,18 +1,0 @@
-import React from 'react';
-
-const Header = () => {
-    return (
-        <header>
-            <h1>Welcome to My Construction Company</h1>
-            <nav>
-                <ul>
-                    <li><a href="#services">Services</a></li>
-                    <li><a href="#projects">Projects</a></li>
-                    <li><a href="#contact">Contact Us</a></li>
-                </ul>
-            </nav>
-        </header>
-    );
-}
-
-export default Header;
