+++ src/components/Process.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const Process = () => {
+    return (
+        <section id="process">
+            <h2>Our Process</h2>
+            <p>It's never in this order</p>
+            <ul>
+                <li>Step 1</li>
+                <li>Step 2</li>
+                <li>Step 3</li>
+            </ul>
+        </section>
+    );
+}
+
+export default Process;
